+++ adept/utils/misc.py
