+++ run.py
@@ -11,7 +11,7 @@ config.update("jax_enable_x64", True)
 import yaml
 
 from adept import ergoExo
-from utils.misc import export_run
+from adept.utils.misc import export_run
 
 
 if __name__ == "__main__":
